added cout the total win, loss, draw by python

+++ convertResult.cpp
@@ -6,7 +6,7 @@ void convertResultsTOXO(){
     for(int total=combinationSize; total>=0 ;total--){
         // cout << "total = " << total << endl;
         for(int oNumber=0;oNumber<=total;oNumber++){
-            unsigned int win=0, draw=0, loss=0;
+            unsigned long long win=0, draw=0, loss=0;
             int xNumber = total - oNumber;
             vector<bool> values(combinations[combinationSize][oNumber] * combinations[(combinationSize-oNumber)][xNumber] * 2);
             readStatesValue(&values, oNumber, xNumber);
